Fixed crash on eye-dropper over colour bar
git-svn-id: svn://svn.xara.com/Trunk/XaraLX@1223 c9fdbc50-1c0f-0410-8134-ac6df33ab1bd

+++ wxOil/ccolbar.cpp
@@ -1305,7 +1305,7 @@ BOOL CColourBar::DoColourPickerColour (wxWindow* colourPicker, wxPoint mousePt, 
 		UINT32 ColourCell = WhereIsMouse (mousePt);
 		static UINT32 lastColourCell = (UINT32)-1;
 
-		if ((ColourCell >= 0) && (lastColourCell != ColourCell))		// we are over a valid colour
+		if ((((INT32)ColourCell) >= 0) && (lastColourCell != ColourCell))		// we are over a valid colour
 		{
 			BOOL validColour = FALSE;
